+++ extra/emailValidaiton/index.js
@@ -1,0 +1,7 @@
+
+const validation = require('./validation').emailvalidaitaion;
+const permission = require('./permissions').permission;
+module.exports = {
+    validation,
+    permission
+}
