+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "successive-training",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "npx babel-node extra"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/satish-agarkar-successive/express-server.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/satish-agarkar-successive/express-server/issues"
+  },
+  "homepage": "https://github.com/satish-agarkar-successive/express-server#readme",
+  "dependencies": {},
+  "devDependencies": {
+    "@babel/core": "^7.6.4",
+    "@babel/node": "^7.6.3",
+    "@babel/preset-env": "^7.6.3"
+  }
+}
